[Silver I] Title: 오르막 수, Time: 0 ms, Memory: 2060 KB -BaekjoonHub

+++ 백준/Silver/11057. 오르막 수/오르막 수.cc
@@ -3,8 +3,8 @@
 using namespace std;
 
 int N;
-long long dp[1001][10] = { 0, };
-long long sol = 0;
+int dp[1001][10] = { 0, };
+int sol = 0;
 
 
 int main() {
